+++ diesel/src/expression/date_and_time.rs
@@ -1,0 +1,35 @@
+use expression::{Expression, SelectableExpression};
+use query_builder::{QueryBuilder, BuildQueryResult};
+use types::{Timestamp, VarChar};
+
+pub struct AtTimeZone<Ts, Tz> {
+    timestamp: Ts,
+    timezone: Tz,
+}
+
+impl<Ts, Tz> AtTimeZone<Ts, Tz> {
+    pub fn new(timestamp: Ts, timezone: Tz) -> Self {
+        AtTimeZone {
+            timestamp: timestamp,
+            timezone: timezone,
+        }
+    }
+}
+
+impl<Ts, Tz> Expression for AtTimeZone<Ts, Tz> where
+    Ts: Expression<SqlType=Timestamp>,
+    Tz: Expression<SqlType=VarChar>,
+{
+    // FIXME: This should be Timestamptz when we support that type
+    type SqlType = Timestamp;
+
+    fn to_sql(&self, out: &mut QueryBuilder) -> BuildQueryResult {
+        try!(self.timestamp.to_sql(out));
+        out.push_sql(" AT TIME ZONE ");
+        self.timezone.to_sql(out)
+    }
+}
+
+impl<Ts, Tz, Qs> SelectableExpression<Qs> for AtTimeZone<Ts, Tz> where
+    AtTimeZone<Ts, Tz>: Expression,
+{}