+++ diesel_tests/tests/types_roundtrip.rs
@@ -9,16 +9,18 @@ pub use diesel::result::Error;
 pub use diesel::data_types::*;
 pub use diesel::types::{NativeSqlType, ToSql, Nullable, Array};
 
-pub fn test_type_round_trips<ST, T>(value: T, type_name: &str) -> bool where
+use diesel::expression::AsExpression;
+
+pub fn test_type_round_trips<ST, T>(value: T) -> bool where
     ST: NativeSqlType,
-    T: ToSql<ST> + Queryable<ST> + PartialEq + ::std::fmt::Debug,
+    T: AsExpression<ST> + Queryable<ST> + PartialEq + Clone + ::std::fmt::Debug,
+    <T as AsExpression<ST>>::Expression: SelectableExpression<()>,
 {
     let connection = connection();
-    let query = format!("SELECT $1::{}", type_name);
-    let result = connection.query_sql_params::<ST, T, ST, T>(&query, &value);
+    let query = select(AsExpression::<ST>::as_expression(value.clone()));
+    let result = query.get_result::<T>(&connection);
     match result {
-        Ok(mut val) => {
-            let res = val.nth(0).unwrap();
+        Ok(res) => {
             if value != res {
                 println!("{:?}, {:?}", value, res);
                 false
@@ -35,25 +37,25 @@ pub fn test_type_round_trips<ST, T>(value: T, type_name: &str) -> bool where
 pub fn id<A>(a: A) -> A { a }
 
 macro_rules! test_round_trip {
-    ($test_name:ident, $sql_type:ident, $tpe:ty, $sql_type_name:expr) => {
-        test_round_trip!($test_name, $sql_type, $tpe, id, $sql_type_name);
+    ($test_name:ident, $sql_type:ident, $tpe:ty) => {
+        test_round_trip!($test_name, $sql_type, $tpe, id);
     };
 
-    ($test_name:ident, $sql_type:ident, $tpe:ty, $map_fn:ident, $sql_type_name:expr) => {
+    ($test_name:ident, $sql_type:ident, $tpe:ty, $map_fn:ident) => {
         #[test]
         fn $test_name() {
             fn round_trip(val: $tpe) -> bool {
-                test_type_round_trips::<types::$sql_type, _>($map_fn(val), $sql_type_name)
+                test_type_round_trips::<types::$sql_type, _>($map_fn(val))
             }
 
             fn option_round_trip(val: Option<$tpe>) -> bool {
                 let val = val.map($map_fn);
-                test_type_round_trips::<Nullable<types::$sql_type>, _>(val, $sql_type_name)
+                test_type_round_trips::<Nullable<types::$sql_type>, _>(val)
             }
 
             fn vec_round_trip(val: Vec<$tpe>) -> bool {
                 let val: Vec<_> = val.into_iter().map($map_fn).collect();
-                test_type_round_trips::<Array<types::$sql_type>, _>(val, concat!($sql_type_name, "[]"))
+                test_type_round_trips::<Array<types::$sql_type>, _>(val)
             }
 
             quickcheck(round_trip as fn($tpe) -> bool);
@@ -63,21 +65,21 @@ macro_rules! test_round_trip {
     }
 }
 
-test_round_trip!(bool_roundtrips, Bool, bool, "boolean");
-test_round_trip!(i16_roundtrips, SmallInt, i16, "int2");
-test_round_trip!(i32_roundtrips, Integer, i32, "int4");
-test_round_trip!(i64_roundtrips, BigInt, i64, "int8");
-test_round_trip!(f32_roundtrips, Float, f32, "real");
-test_round_trip!(f64_roundtrips, Double, f64, "double precision");
-test_round_trip!(string_roundtrips, VarChar, String, "varchar");
-test_round_trip!(text_roundtrips, Text, String, "text");
-test_round_trip!(binary_roundtrips, Binary, Vec<u8>, "bytea");
-test_round_trip!(date_roundtrips, Date, PgDate, "date");
-test_round_trip!(time_roundtrips, Time, PgTime, "time");
-test_round_trip!(timestamp_roundtrips, Timestamp, PgTimestamp, "timestamp");
-test_round_trip!(interval_roundtrips, Interval, PgInterval, "interval");
-test_round_trip!(numeric_roundtrips, Numeric, PgNumeric, "numeric");
-test_round_trip!(naive_datetime_roundtrips, Timestamp, (i64, u32), mk_naive_datetime, "timestamp");
+test_round_trip!(bool_roundtrips, Bool, bool);
+test_round_trip!(i16_roundtrips, SmallInt, i16);
+test_round_trip!(i32_roundtrips, Integer, i32);
+test_round_trip!(i64_roundtrips, BigInt, i64);
+test_round_trip!(f32_roundtrips, Float, f32);
+test_round_trip!(f64_roundtrips, Double, f64);
+test_round_trip!(string_roundtrips, VarChar, String);
+test_round_trip!(text_roundtrips, Text, String);
+test_round_trip!(binary_roundtrips, Binary, Vec<u8>);
+test_round_trip!(date_roundtrips, Date, PgDate);
+test_round_trip!(time_roundtrips, Time, PgTime);
+test_round_trip!(timestamp_roundtrips, Timestamp, PgTimestamp);
+test_round_trip!(interval_roundtrips, Interval, PgInterval);
+test_round_trip!(numeric_roundtrips, Numeric, PgNumeric);
+test_round_trip!(naive_datetime_roundtrips, Timestamp, (i64, u32), mk_naive_datetime);
 
 fn mk_naive_datetime(data: (i64, u32)) -> NaiveDateTime {
     NaiveDateTime::from_timestamp(data.0, data.1 / 1000)
@@ -100,5 +102,5 @@ mod unstable_types {
         time + Duration::new(0, 1) - Duration::new(0, 1)
     }
 
-    test_round_trip!(systemtime_roundtrips, Timestamp, SystemTime, strip_nanosecond_precision, "timestamp");
+    test_round_trip!(systemtime_roundtrips, Timestamp, SystemTime, strip_nanosecond_precision);
 }